requirements.txt
python-telegram-bot[job-queue]==21.11.1
httpx==0.28.1
beautifulsoup4==4.13.4
pytz==2025.2
